refactored routes into sep files, added auth check
